requirements.txt
aiosmtpd
requests
python-dotenv
aiohttp
pytest
pytest-asyncio
aioresponses
sentry-sdk
